src/app/enterprise-security-assistant/requirements.txt
# Enterprise Security Assistant - Python Dependencies
# Python Version: 3.13+

# LangChain and LLM Framework
langchain>=1.2.10
langchain-core>=1.2.14
langchain-openai>=1.1.10
langchain-community>=0.4.1
langgraph>=1.0.9
openai>=1.68.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
redis>=5.0.0
greenlet>=3.1.0

# Vector Database
chromadb>=0.5.0
faiss-cpu>=1.8.0

# Monitoring and Observability
prometheus-client>=0.21.0
opentelemetry-api>=1.28.0
opentelemetry-sdk>=1.28.0
opentelemetry-instrumentation-fastapi>=0.48b0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20

# Utilities
python-dotenv>=1.0.1
loguru>=0.7.3
tenacity>=9.0.0

